+++ application/src/test/java/kitchenpos/application/MenuServiceIntegrateTest.java
@@ -4,11 +4,11 @@ import kitchenpos.execute.ServiceIntegrateTest;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.repository.MenuGroupRepository;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.dto.request.MenuCreateRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
+import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu_group.domain.repository.MenuGroupRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
@@ -61,15 +61,6 @@ class MenuServiceIntegrateTest extends ServiceIntegrateTest {
             // given
             MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.인기_메뉴_생성());
             MenuCreateRequest request = new MenuCreateRequest("치킨", BigDecimal.valueOf(-100), menuGroup.getId());
-
-            // when, then
-            Assertions.assertThrows(IllegalArgumentException.class, () -> menuService.create(request));
-        }
-
-        @Test
-        void 메뉴의_MenuGroupId가_존재하지_않는다면_예외가_발생한다() {
-            // given
-            MenuCreateRequest request = new MenuCreateRequest("치킨", BigDecimal.valueOf(1000), 1L);
 
             // when, then
             Assertions.assertThrows(IllegalArgumentException.class, () -> menuService.create(request));
