+++ menu/src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,8 +1,6 @@
 package kitchenpos.menu.application;
 
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.repository.MenuGroupRepository;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.dto.request.MenuCreateRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
@@ -15,20 +13,14 @@ import java.util.stream.Collectors;
 @Service
 public class MenuService {
     private final MenuRepository menuRepository;
-    private final MenuGroupRepository menuGroupRepository;
 
-    public MenuService(MenuRepository menuRepository, MenuGroupRepository menuGroupRepository) {
+    public MenuService(MenuRepository menuRepository) {
         this.menuRepository = menuRepository;
-        this.menuGroupRepository = menuGroupRepository;
     }
 
     @Transactional
     public Long create(final MenuCreateRequest request) {
-        if (!menuGroupRepository.existsById(request.getMenuGroupId())) {
-            throw new IllegalArgumentException();
-        }
-        MenuGroup menuGroup = menuGroupRepository.getById(request.getMenuGroupId());
-        Menu menu = new Menu(request.getName(), request.getPrice(), menuGroup.getId());
+        Menu menu = new Menu(request.getName(), request.getPrice());
         menu.validatePriceIsNullOrMinus();
         return menuRepository.save(menu).getId();
     }
