+++ menu_group/src/main/java/kitchenpos/menu_group/application/MenuGroupService.java
@@ -1,8 +1,8 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu_group.application;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.repository.MenuGroupRepository;
-import kitchenpos.menu.dto.response.MenuGroupResponse;
+import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu_group.domain.repository.MenuGroupRepository;
+import kitchenpos.menu_group.dto.response.MenuGroupResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
