+++ menu_group/src/main/java/kitchenpos/menu_group/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu_group.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
