+++ menu_group/src/main/java/kitchenpos/menu_group/domain/repository/MenuGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.domain.repository;
+package kitchenpos.menu_group.domain.repository;
 
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu_group.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
