+++ menu_group/src/main/java/kitchenpos/menu_group/domain/repository/MenuProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.domain.repository;
+package kitchenpos.menu_group.domain.repository;
 
 
-import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu_group.domain.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
