+++ menu_group/src/main/java/kitchenpos/menu_group/dto/request/MenuGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.dto.request;
+package kitchenpos.menu_group.dto.request;
 
 public class MenuGroupCreateRequest {
