--- product/src/test/java/execute/ServiceUnitTest.java
@@ -1,14 +1,0 @@
-package execute;
-
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional
-@ExtendWith(MockitoExtension.class)
-@SuppressWarnings("NonAsciiCharacters")
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-public class ServiceUnitTest {
-}
