+++ settings.gradle
@@ -5,3 +5,5 @@ include 'product'
 include 'application'
 include 'common'
 include 'order_table'
+include 'menu_group'
+
